feat: add public project components for enhanced user experience
- Introduced `CopyUrlButton` for easy URL copying.
- Added `ErrorView` to handle project not found scenarios.
- Implemented `PublicKanbanView` and `PublicListView` for task management views.
- Created `KaneoBranding` component for branding consistency.
- Developed `LoadingSkeleton` for loading states.
- Added `ThemeToggle` for user theme preferences.
- Established `task-card` and `task-row` components for task representation.
- Defined `DEFAULT_COLUMNS` for task organization in kanban and list views.
