+++ apps/web/src/components/public-project/copy-url-button.tsx
@@ -1,0 +1,38 @@
+import { Button } from "@/components/ui/button";
+import { Check, Copy } from "lucide-react";
+import { useState } from "react";
+
+export function CopyUrlButton() {
+  const [copied, setCopied] = useState(false);
+
+  const handleCopyUrl = async () => {
+    try {
+      await navigator.clipboard.writeText(window.location.href);
+      setCopied(true);
+      setTimeout(() => setCopied(false), 2000);
+    } catch (error) {
+      console.error("Failed to copy URL:", error);
+    }
+  };
+
+  return (
+    <Button
+      variant="outline"
+      size="sm"
+      onClick={handleCopyUrl}
+      className="gap-2"
+    >
+      {copied ? (
+        <>
+          <Check className="w-4 h-4" />
+          Copied
+        </>
+      ) : (
+        <>
+          <Copy className="w-4 h-4" />
+          Copy URL
+        </>
+      )}
+    </Button>
+  );
+}
